fix: ensure consistent type annotations for performance values in recalculate_score function

+++ app/usecases/performance.py
@@ -148,7 +148,7 @@ def calculate_performances(
             # TODO: report to logserver
             pp = 0.0
         else:
-            pp = round(pp, 3)
+            pp = float(round(pp, 3))
 
         results.append(
             {
